+++ packages/orq-rc/RELEASES.md
@@ -12359,3 +12359,13 @@ Based on:
 - [typescript v4.15.0-rc.47] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.47 - packages/orq-rc
+
+## 2026-09-15 09:37:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.48] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.48 - packages/orq-rc